Made the device list self-updating
Previously, the device list only loaded once on startup.  This was poor
behaviour, as devices can be attached and detached at any time.

Now the attached devices are polled once per second, and the list is
updated accordingly.  The code behind it is a bit messy because the
list model interacts directly with parts of the GUI, but it works
pretty well nevertheless.  There's no mechanism to kick us out of the
configuration screen if the device we're configuring is detached, but
this can be addressed later.  For now, I'm reasonably happy with this
new feature.