Reorganized device list code
Previously, the SelectListStore directly modified the GUI in its
update_items method, which was really just poor form.

This has been corrected by creating a new SelectList to encapsulate all
the GUI elements from the select-stack downward.  This handles showing
the select-none or select-frame elements as appropriate, as well as
adding the list item separators.  It's much nicer than the old way, if I
do say so myself.