Only show Save button when settings are different
Before now, the Save button appeared when the voltage and current
widgets changed their value.  If they were then changed back to the
original values, the Save button was still visible, allowing unnecessary
wear on the flash.  (the sarcastic horror!)

Now, the Save button is only shown when the voltage and current are not
both at their saved values.  A minor tweak, but we're closer to the
original design now.