Error handling when sending commands
Before now, if an error occurred when sending a command, it passed
silently.  This is a Bad Thing™, so it's been fixed.  Now if there's an
error, an error dialog appears with some information about what
happened.  If the error occurred on the Sink page, the program goes back
to the device list after the error dialog is closed.