Return to device list when device is unplugged
There was a bug where a device could end up being programmed with 0
current if the device was disconnected and reconnected while editing its
configuration.  I thought of several solutions for this, but concluded
that the right one would really be to go ahead with my original plan to
kick the user out of the configuration screen when the device was
unplugged.  As usual, the implementation could be better, but it works.