Show empty config when device has no config
Before, we showed the contents of tmpcfg when the device had no
configuration.  This caused some problems when the device was erased,
then opened in the GUI: the contents of tmpcfg were still displayed
despite the device having no configuration, which was confusing.  Now
empty configuration is shown in this case.