Partially implement new UI mockup
The GiveBack flag now has a dedicated row with a switch, rather than a
toggle button the row labeled "Flags".  That GiveBack is one of several
flags seems like an implementation detail from the high level of a GUI.

There is now a combo box to select the dimension of the current-setting
value.  Changing this combo box automatically changes the value to keep
the current the same at the present voltage.  The correct item is
selected when a Sink is chosen from the list, and the unit label always
updates to match the dimension.